% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{clustering}
\alias{clustering}
\title{Hierarchical cluster analysis}
\usage{
clustering(mat, row = TRUE, col = TRUE)
}
\arguments{
\item{mat}{matrix}

\item{row}{boolean values determining if rows should be clustered or hclust object}

\item{col}{boolean values determining if columns should be clustered or hclust object}
}
\value{
matrix of which the order of row and column changed according to clustering
}
\description{
Hierarchical cluster analysis
}
\examples{
test = matrix(rnorm(200), 20, 10)
test[1:10, seq(1, 10, 2)] = test[1:10, seq(1, 10, 2)] + 3
test[11:20, seq(2, 10, 2)] = test[11:20, seq(2, 10, 2)] + 2
test[15:20, seq(2, 10, 2)] = test[15:20, seq(2, 10, 2)] + 4
colnames(test) = paste("Test", 1:10, sep = "")
rownames(test) = paste("Gene", 1:20, sep = "")
clustering(test)
}
