% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{complex}
\alias{complex}
\title{计算复合体的表达值}
\usage{
complex(mtx, gene_name_list)
}
\arguments{
\item{mtx}{表达矩阵，行代表基因，列代表细胞}

\item{gene_name_list}{a list，其中每个元素是亚基的集合}
}
\value{
mtx 将复合体表达值附加到原表达矩阵后的表达矩阵
}
\description{
用复合体的多亚基的几何平均数来表示复合体表达值
}
\details{
为了推断细胞间的相互作用，考虑到配体受体的复合体，需要计算复合体的表达值，便于筛选出特异的相互作用，
将复合体表达值附加到原表达矩阵之后。
}
\examples{
mtx <- matrix(abs(rnorm(200)), 20, 10)
colnames(mtx) = paste("Cell", 1:10, sep = "")
rownames(mtx) = paste("Gene", 1:20, sep = "")
gene_name_list <- list("Gene1",c("Gene2","Gene3"),"Gene4",c("Gene5","Gene6"))
complex(mtx,gene_name_list)

}
