% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{draw_colnames}
\alias{draw_colnames}
\title{根据文字的相对位置布局列名}
\usage{
draw_colnames(coln, position, angle_col, ...)
}
\arguments{
\item{coln}{colnames}

\item{position}{"top" or "bottom"}

\item{angle_col}{旋转角度}

\item{...}{additional arguments passed on to image}
}
\value{
textGrob,字符串的相对位置
}
\description{
根据文字的相对位置布局列名
}
