% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{draw_diagonalmap}
\alias{draw_diagonalmap}
\title{绘制双matrix热图}
\usage{
draw_diagonalmap(leftmatrix, rightmatrix)
}
\arguments{
\item{leftmatrix}{左侧矩阵}

\item{rightmatrix}{右侧矩阵}
}
\value{
双matrix热图
}
\description{
绘制双matrix热图
}
