% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{draw_legend}
\alias{draw_legend}
\title{绘制legend}
\usage{
draw_legend(color, breaks, legend = NA, just = "top", ...)
}
\arguments{
\item{color}{所有颜色}

\item{breaks}{数值区间的边界}

\item{legend}{图例}

\item{just}{"top" or "bottom"}

\item{...}{additional arguments passed on to image}
}
\value{
绘制的legend
}
\description{
绘制legend
}
