% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{draw_matrix}
\alias{draw_matrix}
\title{根据matrix中的颜色绘制成热图}
\usage{
draw_matrix(matrix, border_color = "black")
}
\arguments{
\item{matrix}{颜色矩阵}

\item{border_color}{默认为"black"}
}
\value{
热图
}
\description{
根据matrix中的颜色绘制成热图
}
