% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{draw_rownames}
\alias{draw_rownames}
\title{根据文字的相对位置布局行名}
\usage{
draw_rownames(rown, position, ...)
}
\arguments{
\item{rown}{rownames}

\item{position}{"right" or "left"}

\item{...}{additional arguments passed on to image}
}
\value{
textGrob,字符串的相对位置
}
\description{
根据文字的相对位置布局行名
}
