% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{draw_triangle_down}
\alias{draw_triangle_down}
\title{根据matrix中的颜色绘制成三角热图}
\usage{
draw_triangle_down(matrix, border_color = "grey60", border_size = 0.3)
}
\arguments{
\item{matrix}{表达矩阵}

\item{border_color}{"grey60"}

\item{border_size}{0.3}
}
\value{
三角热图
}
\description{
根据matrix中的颜色绘制成三角热图
}
