% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{find_interactions}
\alias{find_interactions}
\title{Find ligand-receptor pairs from expression matrix}
\usage{
find_interactions(
  object,
  DB,
  celltype = "celltype",
  method = c("expression_threshold", "expression_product", "differential_combinations"),
  threshold = 0,
  product_top = 0.5,
  receiver_celltypes = "ALL",
  sender_celltypes = "ALL",
  marker_filter = FALSE,
  avg_log2FC = 1,
  padj = 0.05
)
}
\arguments{
\item{object}{Seurat object}

\item{DB}{database of ligand-receptor pairs}

\item{celltype}{Seurat object的细胞类型}

\item{method}{"expression_threshold", "expression_product", "differential_combinations"}

\item{threshold}{a number,基因表达均值应高于的阈值}

\item{product_top}{a number,配体受体对的基因表达均值乘积的percentage}

\item{receiver_celltypes}{character,产生receptor的细胞类型}

\item{sender_celltypes}{character,发送ligand的细胞类型}

\item{marker_filter}{logi,True or False}

\item{avg_log2FC}{Limit testing to genes which show, on average, at least X-fold difference (log-scale) between the two groups of cells. Default is 0.25 Increasing logfc.threshold speeds up the function, but can miss weaker signals.}

\item{padj}{Limit testing to genes which show,Default is 0.05}
}
\value{
ligand-receptor pairs from expression matrix
}
\description{
Find ligand-receptor pairs from expression matrix
}
\examples{
hnscc_primary <- readRDS("C:/Users/YWJ/Documents/hnscc_primary.rds")
lrDB <- CellChatDB.human
receiver_celltypes=c("T cell","CAF","Endothelial","Macrophage","myofibroblast")
sender_celltypes="primary"
interaction1 <- find_interactions(object=hnscc_primary,DB=lrDB,celltype="celltype",method="expression_threshold",
                                  receiver_celltypes=receiver_celltypes,
                                  sender_celltypes=sender_celltypes,threshold=3)
interaction2 <- find_interactions(object=hnscc_primary,DB=lrDB,celltype="celltype",method="expression_product",
                                  receiver_celltypes=receiver_celltypes,
                                  sender_celltypes=sender_celltypes,product_top=0.5)
interaction3 <- find_interactions(object=hnscc_primary,DB=lrDB,celltype="celltype",method="differential_combinations",
                                  receiver_celltypes=receiver_celltypes,
                                  sender_celltypes=sender_celltypes,
                                  marker_filter=TRUE,avg_log2FC=1,padj=0.05)

}
