% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{generate_breaks}
\alias{generate_breaks}
\title{根据给定数据上下限，返回区间n个上下限，从小到大排列}
\usage{
generate_breaks(x, n, center = F)
}
\arguments{
\item{x}{data.frame,vector or matrix}

\item{n}{区间数}

\item{center}{使用上下限的最大绝对值替换原有的上下限}
}
\value{
n个区间的上下限
}
\description{
根据给定数据上下限，返回区间n个上下限，从小到大排列
}
