% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{geometric.mean}
\alias{geometric.mean}
\title{Compute the geometric mean}
\usage{
geometric.mean(x, na.rm = TRUE)
}
\arguments{
\item{x}{a numeric vector}

\item{na.rm}{whether remove na}
}
\value{
number,geometric mean
}
\description{
Compute the geometric mean
}
\examples{
x <- 1:10
geometric.mean(x)

}
