% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_expressed_genes}
\alias{get_expressed_genes}
\title{Determine expressed genes of a cell type from a Seurat object single-cell RNA seq dataset}
\usage{
get_expressed_genes(seurat_obj, ident, pct = 0.1)
}
\arguments{
\item{seurat_obj}{Single-cell expression dataset as Seurat object}

\item{ident}{Name of cluster identity/identities of cells}

\item{pct}{We consider genes expressed if they are expressed in at least a specific fraction of cells of a cluster.
This number indicates this fraction. Default: 0.10. Choice of this parameter is important and depends largely on
the used sequencing platform. We recommend to require a lower fraction (like the default 0.10)
for 10X data than for e.g. Smart-seq2 data.}
}
\value{
A character vector with the gene symbols of the expressed genes
}
\description{
\code{get_expressed_genes} Return the genes that are expressed in a given cell cluster based on
the fraction of cells in that cluster that should express the gene.
}
\details{
get_expressed_genes(ident, seurat_obj, pct = 0.10, assay_oi = NULL)
}
\examples{
seuratObj <- readRDS(system.file("extdata","pbmc3k_final.rds", package = "iPhonedev2"))
get_expressed_genes(ident = "CD8 T", seurat_obj = seuratObj, pct = 0.10)

}
