% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{gtable_layout}
\alias{gtable_layout}
\title{根据left、middle和bottom中的行列名、行列数量比例，设置整体布局}
\usage{
gtable_layout(left, middle, bottom, gapsize, fontsize = 7)
}
\arguments{
\item{left}{颜色矩阵}

\item{middle}{颜色矩阵}

\item{bottom}{颜色矩阵}

\item{gapsize}{间隔距离}

\item{fontsize}{字体大小}
}
\value{
根据需求的布局设计
}
\description{
根据left、middle和bottom中的行列名、行列数量比例，设置整体布局
}
