% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{interact_combination}
\alias{interact_combination}
\title{找出配体受体marker基因之间的细胞互作}
\usage{
interact_combination(
  sub_object,
  LR,
  marker_filter = FALSE,
  avg_log2FC = 1,
  padj = 0.05,
  sender_celltypes,
  receiver_celltypes
)
}
\arguments{
\item{sub_object}{seurat object}

\item{LR}{ligand-receptor pairs}

\item{marker_filter}{logi,True or False}

\item{avg_log2FC}{Limit testing to genes which show, on average, at least X-fold difference (log-scale) between the two groups of cells. Default is 0.25 Increasing logfc.threshold speeds up the function, but can miss weaker signals.}

\item{padj}{Limit testing to genes which show,Default is 0.05}

\item{sender_celltypes}{character,发送ligand的细胞类型}

\item{receiver_celltypes}{character,产生receptor的细胞类型}
}
\value{
interaction_combination 配体受体marker基因的细胞互作
}
\description{
sender_celltypes和receiver_celltypes的marker基因的细胞互作
}
\examples{
sub_object = readRDS(system.file("extdata","pbmc3k_final.rds", package = "iPhonedev2"))
LR = Ramilowski_human
sender_celltypes = c("CD8 T","NK")
receiver_celltypes = c("DC","Platelet")
interact_combination(sub_object,LR,marker_filter=FALSE,
avg_log2FC=0.5,padj=0.05,sender_celltypes,receiver_celltypes)

}
