% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{interact_product}
\alias{interact_product}
\title{找出配体受体基因表达乘积top(product_top)的细胞互作}
\usage{
interact_product(
  mtx_lr,
  cell.named.types,
  LR,
  sender_celltypes,
  receiver_celltypes,
  product_top
)
}
\arguments{
\item{mtx_lr}{各细胞类型的基因表达值}

\item{cell.named.types}{由细胞类型组成的向量，向量名为细胞名}

\item{LR}{先验的配体受体对}

\item{sender_celltypes}{character,发送ligand的细胞类型}

\item{receiver_celltypes}{character,产生receptor的细胞类型}

\item{product_top}{a number,percentage}
}
\value{
interaction_product 配体受体基因表达乘积高于某阈值的细胞互作
}
\description{
sender_celltypes和receiver_celltypes的基因表达值均高于设定阈值的细胞互作
}
\examples{
mean_lr = matrix(abs(rnorm(60)), 20, 3)
colnames(mean_lr) = c("Bcell","CAF","Tcell")
rownames(mean_lr) = paste("Gene", 1:20, sep = "")
LR = data.frame(ligand=paste("Gene", 1:10, sep = ""),receptor=paste("Gene", 11:20, sep = ""))
sender_celltypes = c("Bcell","Tcell")
receiver_celltypes = c("CAF","Tcell")
product_top = 0.8
interact_product(mean_lr,LR,sender_celltypes,receiver_celltypes,product_top)

}
