% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{list_mean_lr}
\alias{list_mean_lr}
\title{取出sender和receiver对应的配体受体对的表达值}
\usage{
list_mean_lr(mean_lr, LR, sender_celltypes, receiver_celltypes)
}
\arguments{
\item{mean_lr}{各细胞类型的基因表达值}

\item{LR}{先验的配体受体对}

\item{sender_celltypes}{character,发送ligand的细胞类型}

\item{receiver_celltypes}{character,产生receptor的细胞类型}
}
\value{
list_mean_lr a list,sender的配体表达值，receiver的受体表达值，顺序与LR一致
}
\description{
取出sender和receiver对应的配体受体对的表达值
}
\examples{
mtx <- matrix(abs(rnorm(60)), 20, 3)
colnames(mtx) = c("Bcell","CAF","Tcell")
rownames(mtx) = paste("Gene", 1:20, sep = "")
LR = data.frame(ligand=paste("Gene", 1:10, sep = ""),receptor=paste("Gene", 11:20, sep = ""))
sender_celltypes = "Bcell"
receiver_celltypes = c("CAF","Tcell")
list_mean_lr(mean_lr,LR,sender_celltypes,receiver_celltypes)

}
