% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sig.R
\name{logfc}
\alias{logfc}
\title{计算ligand和receptor在指定细胞类型中的logfc}
\usage{
logfc(object, siglr)
}
\arguments{
\item{object}{Seurat object}

\item{siglr}{data.frame,必须包含ligand,receptor,from,to，pval}
}
\value{
list,receptor_logfc,receptor基因在特定细胞类型中的logfc值;
ligand_logfc,ligand基因在特定细胞类型中的logfc值
}
\description{
计算ligand和receptor在指定细胞类型中的logfc
}
\examples{
load(url("https://figshare.com/ndownloader/files/37663929"))
interaction <- data.frame(ligand = c("MDK","LAMB3"),
                          receptor = c("NCL","CD44"),
                          from = "primary",to = "Macrophage",
                          pval = c(0,0))
logfc <- function(object=seurat,siglr=interaction)

}
