% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sig2.R
\name{logfc2}
\alias{logfc2}
\title{计算ligand和receptor在指定细胞类型中的logfc}
\usage{
logfc2(object1, object2, sig1, sig2, sample1 = "sample1", sample2 = "sample2")
}
\arguments{
\item{object1}{Seurat object1}

\item{object2}{Seurat object2}

\item{sig1}{significant result of object1}

\item{sig2}{significant result of object2}

\item{sample1}{name of sample1}

\item{sample2}{name of sample2}
}
\value{
list,receptor_logfc,receptor基因在特定细胞类型中的logfc值;
ligand_logfc,ligand基因在特定细胞类型中的logfc值
}
\description{
计算ligand和receptor在指定细胞类型中的logfc
}
