% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{lr2map2}
\alias{lr2map2}
\title{绘制热图}
\usage{
lr2map2(
  left.mat,
  up.middle.mat,
  down.middle.mat,
  bottom.mat,
  left.colors = NA,
  middle.colors = NA,
  bottom.colors = NA,
  levels.mat = NA,
  ...
)
}
\arguments{
\item{left.mat}{left矩阵}

\item{up.middle.mat}{up.middle矩阵}

\item{down.middle.mat}{down.middle矩阵}

\item{bottom.mat}{bottom矩阵}

\item{left.colors}{left矩阵颜色}

\item{middle.colors}{middle矩阵颜色}

\item{bottom.colors}{bottom矩阵颜色}

\item{levels.mat}{character matrix的level}

\item{...}{additional arguments passed on to image}
}
\value{
热图
}
\description{
绘制热图
}
