% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{map}
\alias{map}
\title{绘制细胞互作热图}
\usage{
map(res, levels.mat = c("0", "1"), transformation = F)
}
\arguments{
\item{res}{list,细胞互作分析结果，包含pvalue,significant,receptor_logfc,ligand_logfc}

\item{levels.mat}{character matrix的level}

\item{transformation}{数据变换，把极值转换到特定范围内}
}
\value{
细胞互作热图
}
\description{
绘制细胞互作热图
}
