% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mean_celltype}
\alias{mean_celltype}
\title{计算细胞类型的基因表达均值}
\usage{
mean_celltype(mtx, cell.named.types)
}
\arguments{
\item{mtx}{表达矩阵，行代表基因，列代表细胞}

\item{cell.named.types}{由细胞类型组成的向量，向量名为细胞名}
}
\value{
mean_lr 各细胞类型的基因表达均值
}
\description{
计算细胞类型的基因表达均值
}
\examples{
mtx <- matrix(abs(rnorm(240)), 20, 12)
colnames(mtx) = paste("Cell", 1:12, sep = "")
rownames(mtx) = paste("Gene", 1:20, sep = "")
cell.named.types = rep(c("Bcell","CAF","Tcell"),4)
names(cell.named.types) = colnames(mtx)
mean_celltype(mtx,cell.named.types)

}
