% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sig.R
\name{permutation}
\alias{permutation}
\title{针对两种细胞类型中的某一ligand-receptor pairs，通过随机抽样获得100个ligand-receptor表达均值的乘积，
用于后面permutation单边检验}
\usage{
permutation(
  seurat,
  l,
  r,
  n_sender,
  n_receiver,
  iteration = 100,
  method = c("product", "sum")
)
}
\arguments{
\item{seurat}{Seurat object}

\item{l}{ligand基因名}

\item{r}{receptor基因名}

\item{n_sender}{sender_celltype的细胞数目}

\item{n_receiver}{receiver_celltype的细胞数目}

\item{iteration}{迭代次数，默认值为100}

\item{method}{"product" or "sum"}
}
\value{
vector,其中的每一个值都是permutation得到的两个mean值的乘积
}
\description{
针对两种细胞类型中的某一ligand-receptor pairs，通过随机抽样获得100个ligand-receptor表达均值的乘积，
用于后面permutation单边检验
}
\examples{
load(url("https://figshare.com/ndownloader/files/37663929"))
l <- "TIGIT"
r <- "PVR"
n_sender <- 50
n_receiver <- 80
permutation(seurat,l,r,n_sender,n_receiver,iteration=100,method="product")

}
