% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{production}
\alias{production}
\title{计算矩阵乘积}
\usage{
production(vector1, vector2)
}
\arguments{
\item{vector1}{某一ligand在sender_celltype的表达均值}

\item{vector2}{某一receptor在receiver的表达均值}
}
\value{
vector1和vector1的乘积
}
\description{
计算矩阵乘积
}
