% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{set_height_width}
\alias{set_height_width}
\title{根据一组文本字符串长度，设置其水平或竖直时的宽度或高度}
\usage{
set_height_width(texts, hw, angle_col = NULL, addtion = 5, fontsize = 7)
}
\arguments{
\item{texts}{vector文本}

\item{hw}{"height" or "width"}

\item{angle_col}{角度}

\item{addtion}{空白宽度或高度}

\item{fontsize}{字体大小}
}
\value{
最长的字符串的高度或宽度
}
\description{
根据一组文本字符串长度，设置其水平或竖直时的宽度或高度
}
