% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sig2.R
\name{sig2}
\alias{sig2}
\title{利用permutation计算配体受体对的pval,找出有显著意义的配体受体对}
\usage{
sig2(
  object1,
  object2,
  interaction1,
  interaction2,
  pval = 0.05,
  method = "product",
  sample1 = "sample1",
  sample2 = "sample2"
)
}
\arguments{
\item{object1}{Seurat object1}

\item{object2}{Seurat object2}

\item{interaction1}{data.frame,必须包含ligand,receptor,from,to}

\item{interaction2}{data.frame,必须包含ligand,receptor,from,to}

\item{pval}{number,默认值为0.05}

\item{method}{"product" or "sum"}

\item{sample1}{name of sample1}

\item{sample2}{name of sample2}
}
\value{
list,pvalue,有显著意义的配体受体对;
significant,data.frame,行是ligand基因名,列是receptor基因名;
receptor_logfc,receptor基因在特定细胞类型中的logfc值;
ligand_logfc,ligand基因在特定细胞类型中的logfc值
}
\description{
利用permutation计算配体受体对的pval,找出有显著意义的配体受体对
}
