% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Single.CellSignalR_human}
\alias{Single.CellSignalR_human}
\title{Ligand-receptor interactions in CellChat database for human}
\format{
A list includes the ligand-receptor interactions
}
\usage{
Single.CellSignalR_human
}
\description{
The ligand-receptor interaction database curated in CellChat tool
}
\keyword{datasets}
