% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{transformation}
\alias{transformation}
\title{数据变换，把极值转换到特定范围内}
\usage{
transformation(mat, range = c(-1, 1))
}
\arguments{
\item{mat}{data.frame or matrix}

\item{range}{转换的范围}
}
\value{
把极值转换到特定范围内的数据
}
\description{
数据变换，把极值转换到特定范围内
}
