% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{value_vec_colours}
\alias{value_vec_colours}
\title{根据颜色和数值区间，把数值向量置换为颜色向量}
\usage{
value_vec_colours(x, col = hue_pal()(20), breaks, na_col = "white")
}
\arguments{
\item{x}{numeric vector}

\item{col}{colours}

\item{breaks}{数值区间}

\item{na_col}{默认值，"white"}
}
\value{
颜色向量
}
\description{
根据颜色和数值区间，把数值向量置换为颜色向量
}
