% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{values2colors}
\alias{values2colors}
\title{将matrix中的数值转换成颜色}
\usage{
values2colors(mat, col = hue_pal()(20), na_col = "white", levels.mat = NA)
}
\arguments{
\item{mat}{matrix}

\item{col}{colours}

\item{na_col}{默认为"white"}

\item{levels.mat}{character matrix的level}
}
\value{
list,breaks为数值区间，col.mat为颜色矩阵
}
\description{
将matrix中的数值转换成颜色
}
